Adding log-emssage and warn-user actions. Adding conditions: not, regex, disabled. Adding a condition which checks against a list of a given set of words. Added list of bad words in all languages

+++ src/rules/actions/delete-message.js
@@ -7,8 +7,23 @@ class DeleteMessageAction extends Action {
         this.slackApi = this.context.slackApis[rule.user];
     }
     async performAction(context, message) {
-        await this.slackApi.deleteMessage(message.channel, message.ts);
-        context.logger.info({to: message.user, rule: context.rule.name}, "Deleted Message");
+        try {
+            if (!message.ts || !message.channel || !message.user) {
+                context.logger.warn({
+                    to: message.user,
+                    botId: message.bot_id,
+                    rule: context.rule.name,
+                    channel: message.channel,
+                    ts: message.ts
+                }, "Cant delete message that doesn't exist.");
+                return;
+            }
+            await this.slackApi.deleteMessage(message.channel, message.ts);
+            context.logger.info({to: message.user, rule: context.rule.name}, "Deleted Message");
+        } catch (err) {
+            context.logger.error({err, to: message.user, rule: context.rule.name}, "Cant delete message.");
+        }
+
     }
 }
